Updated App and its CSS
