--- client/src/App.js
@@ -1,29 +1,0 @@
-import './App.css';
-import { useEffect, useState } from 'react';
-
-function App() {
-  const [message, setMessage] = useState('');
-
-  useEffect(() => {
-    fetch("/api/hello")
-      .then((response) => response.json())
-      .then((data) => {
-        setMessage(data.message);
-      })
-      .catch((error) => {
-        console.error("Error fetching data:", error);
-      });
-  }, []);
-
-  return (
-    <div className="App">
-      <header className="App-header">
-        <p>
-          {message && <span> Server says: {message}</span>}
-        </p>
-      </header>
-    </div>
-  );
-}
-
-export default App;
